Refactor callback handler
