+++ pkg/keycloak/proxy/oauth_proxy.go
@@ -71,27 +71,26 @@ type RequestScope struct {
 	Logger  *zap.Logger
 }
 
+type RealmRoles struct {
+	Roles []string `json:"roles"`
+}
+
+// Extract custom claims
+type custClaims struct {
+	Email          string                    `json:"email"`
+	PrefName       string                    `json:"preferred_username"`
+	RealmAccess    RealmRoles                `json:"realm_access"`
+	Groups         []string                  `json:"groups"`
+	ResourceAccess map[string]interface{}    `json:"resource_access"`
+	FamilyName     string                    `json:"family_name"`
+	GivenName      string                    `json:"given_name"`
+	Username       string                    `json:"username"`
+	Authorization  authorization.Permissions `json:"authorization"`
+}
+
 // ExtractIdentity parse the jwt token and extracts the various elements is order to construct
 func ExtractIdentity(token *jwt.JSONWebToken) (*UserContext, error) {
 	stdClaims := &jwt.Claims{}
-
-	type RealmRoles struct {
-		Roles []string `json:"roles"`
-	}
-
-	// Extract custom claims
-	type custClaims struct {
-		Email          string                    `json:"email"`
-		PrefName       string                    `json:"preferred_username"`
-		RealmAccess    RealmRoles                `json:"realm_access"`
-		Groups         []string                  `json:"groups"`
-		ResourceAccess map[string]interface{}    `json:"resource_access"`
-		FamilyName     string                    `json:"family_name"`
-		GivenName      string                    `json:"given_name"`
-		Username       string                    `json:"username"`
-		Authorization  authorization.Permissions `json:"authorization"`
-	}
-
 	customClaims := custClaims{}
 
 	err := token.UnsafeClaimsWithoutVerification(stdClaims, &customClaims)
